+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add `takeWhileEnd` and `dropWhileEnd`.
 * Add `count`.
 * Add an optimized `split` function.
+* Add `splitInit`.
 
 ## 0.1.3.0 -- 2019-09-15