+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `count`.
 * Add an optimized `split` function.
 * Add `splitInit`.
+* Add `copy`.
 
 ## 0.1.3.0 -- 2019-09-15