+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add an optimized `split` function.
 * Add `splitInit`.
 * Add `copy`.
+* Add `pin`.
 
 ## 0.1.3.0 -- 2019-09-15